requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@30eff605df5db9ba20289a0c7471e4d8547a5a8f
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.8
anyio==3.7.0
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bitarray==2.8.1
bitsandbytes==0.39.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.0
datasets==2.13.0
decorator==5.1.1
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.1
exceptiongroup==1.1.1
executing==1.2.0
-e git+https://github.com/pytorch/fairseq@100cd91db19bb27277a06a25eb4154c805b10189#egg=fairseq
fastapi==0.97.0
ffmpy==0.3.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
gradio==3.35.2
gradio_client==0.2.7
grpcio==1.51.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
hydra-core==1.0.7
idna==3.4
importlib-resources==5.12.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.2
lit==16.0.6
loralib==0.1.1
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.0
omegaconf==2.0.6
orjson==3.9.1
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.6.0
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
ray==2.5.0
regex==2023.6.3
requests==2.31.0
sacrebleu==2.3.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.25.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenize-rt==5.1.0
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@881c0df952fbe1e166033806f9128b3c52dc507d
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.3
uvicorn==0.22.0
wandb==0.15.4
wcwidth==0.2.6
websockets==11.0.3
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
